refactor: 拆分文章页独立职责
