+++ app/src/androidTest/java/com/github/zly2006/zhihu/CollectionScreenInstrumentedTest.kt
@@ -25,11 +25,11 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performScrollToNode
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.zly2006.zhihu.data.Collection
 import com.github.zly2006.zhihu.navigation.CollectionContent
 import com.github.zly2006.zhihu.test.MainActivityComposeRule
 import com.github.zly2006.zhihu.test.resetAppPreferences
 import com.github.zly2006.zhihu.test.setScreenContent
-import com.github.zly2006.zhihu.ui.Collection
 import com.github.zly2006.zhihu.ui.CollectionScreen
 import org.junit.Assert.assertEquals
 import org.junit.Before
