+++ shared/src/commonMain/kotlin/com/github/zly2006/zhihu/data/Collection.kt
@@ -1,0 +1,51 @@
+/*
+ * Zhihu++ - Free & Ad-Free Zhihu client for all platforms.
+ * Copyright (C) 2024-2026, zly2006 <i@zly2006.me>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation (version 3 only).
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.github.zly2006.zhihu.data
+
+import com.github.zly2006.zhihu.shared.data.Person
+import com.github.zly2006.zhihu.shared.data.ZhihuPaging
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Collection(
+    val id: String,
+    val isFavorited: Boolean = false,
+    val type: String = "collection",
+    val title: String = "",
+    val isPublic: Boolean = false,
+    val url: String = "",
+    val description: String = "",
+    val followerCount: Int = 0,
+    val answerCount: Int = 0,
+    val itemCount: Int = 0,
+    val likeCount: Int = 0,
+    val viewCount: Int = 0,
+    val commentCount: Int = 0,
+    val isFollowing: Boolean = false,
+    val isLiking: Boolean = false,
+    val createdTime: Long = 0L,
+    val updatedTime: Long = 0L,
+    val creator: Person? = null,
+    val isDefault: Boolean = false,
+)
+
+@Serializable
+data class CollectionResponse(
+    val data: List<Collection>,
+    val paging: ZhihuPaging,
+)
