+++ shared/src/commonMain/kotlin/com/github/zly2006/zhihu/ui/components/CollectionDialogComponent.kt
@@ -50,7 +50,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import com.github.zly2006.zhihu.ui.Collection
+import com.github.zly2006.zhihu.data.Collection
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
